+++ src/main/java/com/github/jnhyperion/hyperrobotframeworkplugin/psi/element/DefinedKeyword.java
@@ -2,11 +2,15 @@ package com.github.jnhyperion.hyperrobotframeworkplugin.psi.element;
 
 import com.intellij.psi.PsiElement;
 
+import java.util.Collection;
+
 public interface DefinedKeyword {
 
    String getKeywordName();
 
-   boolean hasArguments();
+   boolean hasParameters();
+
+   Collection<DefinedParameter> getParameters();
 
    boolean matches(String text);