+++ src/main/java/com/github/jnhyperion/hyperrobotframeworkplugin/psi/element/KeywordInvokableImpl.java
@@ -4,15 +4,26 @@ import com.github.jnhyperion.hyperrobotframeworkplugin.psi.ref.RobotKeywordRefer
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiReference;
+import org.jetbrains.annotations.NotNull;
+
 import java.util.Collection;
 import java.util.Collections;
-
-import org.jetbrains.annotations.NotNull;
 
 public class KeywordInvokableImpl extends RobotPsiElementBase implements KeywordInvokable {
 
    public KeywordInvokableImpl(@NotNull ASTNode node) {
       super(node);
+   }
+
+   @NotNull
+   @Override
+   public Collection<Parameter> getParameters() {
+      PsiElement parent = this.getParent();
+      if (parent instanceof KeywordStatement) {
+         return ((KeywordStatement) parent).getParameters();
+      } else {
+         return Collections.emptySet();
+      }
    }
 
    @NotNull